requirements.txt
numpy~=1.24.3
mediapipe~=0.10.7
python-vlc~=3.0.20123
opencv-python~=4.8.1.78